pyproject.toml
[tool.poetry]
name = "luna"
version = "0.1.0"
description = ""
authors = ["Kristian <kristianmischke@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
transformers = "^4.8.2"
torch = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
